+++ app/src/main/java/io/novafoundation/nova/app/root/navigation/swap/SwapNavigator.kt
@@ -3,7 +3,7 @@ package io.novafoundation.nova.app.root.navigation.swap
 import io.novafoundation.nova.app.R
 import io.novafoundation.nova.app.root.navigation.BaseNavigator
 import io.novafoundation.nova.app.root.navigation.NavigationHolder
-import io.novafoundation.nova.feature_swap_api.presentation.SwapRouter
+import io.novafoundation.nova.feature_swap_impl.presentation.SwapRouter
 
 class SwapNavigator(
     private val navigationHolder: NavigationHolder
