+++ feature-assets/src/main/java/io/novafoundation/nova/feature_assets/presentation/balance/list/di/BalanceListModule.kt
@@ -67,7 +67,7 @@ class BalanceListModule {
         walletInteractor: WalletInteractor,
         assetsListInteractor: AssetsListInteractor,
         selectedAccountUseCase: SelectedAccountUseCase,
-        assetsRouter: AssetsRouter,
+        router: AssetsRouter,
         currencyInteractor: CurrencyInteractor,
         balanceBreakdownInteractor: BalanceBreakdownInteractor,
         externalBalancesInteractor: ExternalBalancesInteractor,
@@ -78,7 +78,7 @@ class BalanceListModule {
             walletInteractor = walletInteractor,
             assetsListInteractor = assetsListInteractor,
             selectedAccountUseCase = selectedAccountUseCase,
-            assetsRouter = assetsRouter,
+            router = router,
             currencyInteractor = currencyInteractor,
             balanceBreakdownInteractor = balanceBreakdownInteractor,
             externalBalancesInteractor = externalBalancesInteractor,
