+++ feature-swap-impl/src/main/java/io/novafoundation/nova/feature_swap_impl/di/SwapFeatureComponent.kt
@@ -7,7 +7,7 @@ import io.novafoundation.nova.common.di.scope.FeatureScope
 import io.novafoundation.nova.core_db.di.DbApi
 import io.novafoundation.nova.feature_account_api.di.AccountFeatureApi
 import io.novafoundation.nova.feature_swap_api.di.SwapFeatureApi
-import io.novafoundation.nova.feature_swap_api.presentation.SwapRouter
+import io.novafoundation.nova.feature_swap_impl.presentation.SwapRouter
 import io.novafoundation.nova.feature_swap_impl.presentation.confirmation.di.SwapConfirmationComponent
 import io.novafoundation.nova.feature_swap_impl.presentation.main.di.SwapMainSettingsComponent
 import io.novafoundation.nova.feature_wallet_api.di.WalletFeatureApi
