+++ feature-swap-impl/src/main/java/io/novafoundation/nova/feature_swap_impl/presentation/SwapRouter.kt
@@ -1,4 +1,4 @@
-package io.novafoundation.nova.feature_swap_api.presentation
+package io.novafoundation.nova.feature_swap_impl.presentation
 
 import io.novafoundation.nova.common.navigation.ReturnableRouter
